Clean up MSVC warnings
Use explicit casts around SQLite int-sized APIs and generated indices, avoid unreachable optional narrowing paths, and remove an unused caught exception name.

+++ include/zxorm/common.hpp
@@ -66,7 +66,7 @@ namespace zxorm {
             const auto it = std::find(a.begin(), a.end(), true);
 
             if (it == a.end()) return -1;
-            return std::distance(a.begin(), it);
+            return static_cast<int>(std::distance(a.begin(), it));
         }
 
         public:
